3/4/16
Created an NPC class for Non-Player Characters. Also created a TextBox
class.

When the player interacts with an NPC, it will display a TextBox with a
list of text that the NPC will say. When the player clicks a button, it
will go to the next item of text.

TODO:
- Actually draw the NPCs on screen.
- Create a bool for when the player is interacting with the NPC.
- Take the test NPC out of the HUDManager class. Instead we should have
a list of NPCs in the World class, for example, and update and draw
them there.