Revert to inheritance system
I still believe the previous item system is more extensible, more
organized, and ultimately simpler…

When implementing the XML files, I realized some of the drawbacks of
the other system. I don’t think having mutable items (where one apple
should be equivalent to every other apple but actually doesn’t have to
be) is best. Plus, having to deal with items as a pseudo-value type
(when making a new item, you have to duplicate it using a shallow copy
and not, as in the previous system, an intuitive constructor).

I’ll work on incorporating some of the other changes from the merge.