Better organization of World constructor, combined PlayerInput class with Player class
Additionally, removed extraneous references to world in entities. I
don’t think the references should be stored, as this is the whole
purpose of encapsulation and methods. Instead, if a particular method
needs access to the world, it should gain it as one of its arguments
(this is what we’ve been doing with the Update methods thus far). This
is a bit more failsafe, because you know that, e.g., the only time a
Player can modify or access the world is when it is explicitly passed
to it as an argument of a function.

Having references to the world inside each class is effectively like
making World a static public class or a singleton class—so far, we’re
straying away from this design pattern. This again will make it easier
when we have to change worlds (when moving onto a new level)—the
player’s data is independent from the world’s, and when the player (or
any other entity) wants access to it, it gains it only temporarily.