Added more enemies + fixed issue with floor items
Issue: some of the items picked off of the floor had not had their
quantity set yet so it remained at zero. As soon as the player picked
it up, it would disappear from the inventory because it had a quantity
of zero. Now all of the quantities have been set.