Version 1.0 of the game is basically done
The game is basically done now. It is playable, and has pretty basic
story progression. There are enemies and bosses, item drops, etc. There
is still a lot of work to do to make it better though, but it can still
be played and enjoyed as it is now.

TODO:
- Add more story elements
- More items and weapons
- Possibly more floors to explore / something to do after beating the
boss
- Readjust weapon damages, chest spawns, and enemy item drops.
- Optimization to improve performance
- Multiplayer mode so that groups of people can explore the caves and
fight bosses together
- A save feature so that you don’t have to do it all in one run.
- Better enemy AI, including enemies that throw projectiles
- Animations
- Probably many other features that I forgetting right now